refactor: move topic feed logic into the `FeedController`
