+++ lib/changelog_web/router.ex
@@ -148,14 +148,9 @@ defmodule ChangelogWeb.Router do
     get "/:slug/feed", FeedController, :podcast
     get "/plusplus/:slug/feed", FeedController, :plusplus
     get "/metacast/:slug/feed", FeedController, :metacast
-  end
-
-  scope "/", ChangelogWeb do
-    pipe_through [:feed]
-
-    get "/topic/:slug/feed", TopicFeedController, :topic
-    get "/topic/:slug/news/feed", TopicFeedController, :news
-    get "/topic/:slug/podcasts/feed", TopicFeedController, :podcasts
+    get "/topic/:slug/feed", FeedController, :topic
+    get "/topic/:slug/news/feed", FeedController, :topic_news
+    get "/topic/:slug/podcasts/feed", FeedController, :topic_podcasts
   end
 
   scope "/", ChangelogWeb do
