--- lib/changelog_web/views/topic_feed_view.ex
@@ -1,11 +1,0 @@
-defmodule ChangelogWeb.TopicFeedView do
-  use ChangelogWeb, :public_view
-
-  alias Changelog.{}
-
-  alias ChangelogWeb.{}
-
-  def render(_conn, _params) do
-    # TODO
-  end
-end
