--- test/changelog_web/views/topic_feed_view_test.exs
@@ -1,13 +1,0 @@
-defmodule ChangelogWeb.TopicFeedViewTest do
-  use ChangelogWeb.ConnCase, async: true
-
-  # alias ChangelogWeb.FeedView
-
-  topic "" do
-
-  end
-
-  topic "" do
-
-  end
-end
